backend/requirements.txt
asgiref==3.3.1
astroid==2.4.2
autopep8==1.5.5
certifi==2020.12.5
chardet==4.0.0
Django==3.1.6
django-filter==2.4.0
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
idna==2.10
importlib-metadata==3.4.0
isort==5.7.0
lazy-object-proxy==1.4.3
Markdown==3.3.3
mccabe==0.6.1
psycopg2==2.8.6
pycodestyle==2.6.0
PyJWT==2.0.1
pytz==2021.1
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
typing-extensions==3.7.4.3
urllib3==1.26.3
wrapt==1.12.1
zipp==3.4.0
